+++ templates/zerver/api/changelog.md
@@ -10,6 +10,11 @@ below features are supported.
 
 ## Changes in Zulip 2.2
 
+**Feature level 12**
+
+* [`GET users/{user_id}/subscription/{stream_id}`](/api/get-subscription-status):
+  Endpoint added.
+
 **Feature level 11**
 
 * [`POST /register`](/api/register-queue): Added